[Databend] Refactor tests
