+++ .github/workflows/main.yml
@@ -197,7 +197,9 @@ jobs:
         run: mvn -B package -DskipTests=true
       - name: Run Tests
         run: |
-          DATABEND_AVAILABLE=true mvn -Dtest=TestDatabend test
+          DATABEND_AVAILABLE=true mvn -Dtest=TestDatabendTLP test
+          DATABEND_AVAILABLE=true mvn -Dtest=TestDatabendNoREC test
+          DATABEND_AVAILABLE=true mvn -Dtest=TestDatabendPQS test
 
   duckdb:
     name: DBMS Tests (DuckDB)
